requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
dnspython==1.15.0
https://github.com/eventlet/eventlet/archive/339c7ad4eaa4766ac8135fd13eed3adeaa43b276.zip
Flask==1.0.2
Flask-SocketIO==3.0.1
greenlet==0.4.13
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
monotonic==1.5
pkg-resources==0.0.0
PyJWT==1.6.4
pymongo==3.6.1
pyTelegramBotAPI==3.6.3
python-engineio==2.1.1
python-socketio==1.9.0
requests==2.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
